requirements.txt
#boto3==1.3.1
boto==2.41.0
git+https://github.com/quarrius/pyfilesystem.git@quarrius
-r hammer/toybox/requirements.txt
